pyproject.toml
[project]
name = "localtinyimg"
version = "0.1.0"
description = ""
authors = [
    {name = "janes128",email = "jimmy851208@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit (>=1.44.1,<2.0.0)",
    "pillow (>=11.2.1,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
